Added note details and database functionality
I created the basic structure including the model and note categories page (with minimal functionality which I used to pass a hardcoded note header object to call the note details page). The note details page will query and display notes, and will include add/edit/delete functions for new note (title and note) that will be added later on.
Whoever is working on the note categories or courses page would include an on click function to navigate to the note detail page by passing the parent Note object.
So there are two tables in sqfliet ---> The notes table and the notedetails table. So far testing has been done with passed in hard-coded values that need to be saved and then have the program re-run to see the changes but that can be worked on later. I think this is good enough for the project check-in.

I will continue working on the note detail page but thought I should share this structure so we don't go off in different directions. Mainly everyone needs to understand and look at the note.dart file which has model and functions that will be needed for the note categories or courses page as well.

I added another main(2).dart as that is what I used to be able to start testing the notes detail page. Now that we got the login page with firebase, with the login APi that was chosen we just need to to code the GUI for it. Also, we will have to combine the two main.dart files to accommodate both the firebase login and the note details page. 
Note: When someone pulls this project as of now, you would have to change the package name in main(2).dart to whatever folder you pulled the project into. I used import 'package:notepad/notes.dart'; as I just put an arbirtrary name of notepad for when I was testing. Also the sqflite and path-provider dependencies need to be added as well for the notes.dart file but those can be quickly added through quick fix.